requirements.txt
Flask==3.1.2
boto3==1.40.54
python-dotenv==1.1.1
Werkzeug==3.1.3
Jinja2==3.1.6
itsdangerous==2.2.0